fix: rename journey data file so it no longer collides by case
ProjectJourney.tsx and projectJourney.ts differed only in the case of
the first letter, which is ambiguous on case-insensitive filesystems
(Windows, macOS) and made the type checker resolve the component import
to the data module. Renamed the data module to projectJourneyData.ts so
the two names are clearly distinct, and updated the import.
